a more efficient FEM nodes system
