--- projects/FEM/testopensim.cpp
@@ -1,1 +1,0 @@
-#include <OpenSim/OpenSim.h>
